logistics_backend/requirements.txt
aiogram==2.22.2
aiohttp==3.8.3
aiosignal==1.2.0
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.7.2
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
beautifulsoup4==4.12.2
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.6
distro==1.9.0
Django==4.2.10
django-cors-headers==4.8.0
djangorestframework==3.15.1
exceptiongroup==1.2.2
flake8==5.0.4
flake8-broken-line==0.6.0
flake8-isort==6.0.0
flake8-plugin-utils==1.3.2
flake8-return==1.2.0
frozenlist==1.3.1
greenlet==2.0.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.4
isort==5.11.4
jiter==0.7.0
mccabe==0.7.0
multidict==6.0.2
mypy==0.991
mypy-extensions==0.4.3
openai==1.53.0
pep8-naming==0.13.3
pequests==0.0.1
psycopg2==2.9.10
psycopg2-binary==2.9.7
pycodestyle==2.9.1
pydantic==2.9.2
pydantic_core==2.23.4
pyflakes==2.5.0
pytz==2022.5
requests==2.28.2
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.20
sqlparse==0.4.4
tomli==2.0.1
tqdm==4.66.6
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.14
yarl==1.8.1
